requirements.txt
facebook_business
gspread
google-auth
pandas
numpy
